Update code to working
Cleaned up the urls.py to not start with a ^, which was causing issues
when generating the URLs for django 3.2

Updated the models.py and jwt.py to no longer use the protected _params
attribute that wasn't defined, and instead accessing the member for alg
directly.

Updated the README to include the 'modeltranslation' installed
application and updating the URLs that need to be defined.